zut je m etais trompé
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@9802 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _test_/podcast_client/podcast_client.php
@@ -22,28 +22,8 @@
 	
 		list($le_lien, $id_syndic, $data) = func_get_arg(0);
 		//print_r($data);
-		
-		
-		//trouver un flash dailymotion
-		# <media:content url="http://www.dailymotion.com/swf/3ndb67rMbTuLh8E2i" type="application/x-shockwave-flash" duration="520" width="320" height="240"/>
-		//echo $item ;
-		if (preg_match(',(<media:content[^>]*type="video/x-flv".*>),i',
-		$data['item'], $match)) {
-			$go=str_replace('media:content','mediacontent',$match[1]);
-			
-			$data['enc_flash']['url'] = extraire_attribut($go, 'url');
-			$data['enc_flash']['duration'] = extraire_attribut($go, 'duration');
-			$data['enc_flash']['width'] = extraire_attribut($go, 'width');
-			$data['enc_flash']['height'] = extraire_attribut($go, 'height');
-			$data['enc_flash']['type'] = trim(extraire_attribut($go,'type'));
-			if(!$data['enc_flash']['type']) $data['enc_flash']['type'] = substr($data['enc_flash']['url'], -3);
-			//print_r($data['flash']);
-			
-			//$data['url_flash'] = str_replace('&amp;', '&',
-			//	trim(extraire_attribut($match[1], 'url')));
-		}
-		
 		//mettre les documents dans un tableau
+		
 		$enclosure = $data['enclosures'] ;
 		// href et type sont obligatoires
 			if (($enc_regs_url = extraire_attribut($enclosure,'href')
@@ -55,19 +35,32 @@
 				if(!$data['enc_enclosure']['type']) $data['enc_enclosure']['type'] = substr($url, -3);
 			}
 		
-
+		// trouver une image dans le descriptif (flickr)
 		$enclosure = $data['descriptif'] ;
-		// trouver une image dans le descriptif
 			if (($enc_regs_url = extraire_attribut($enclosure,'src'))) {
 				$url  = substr(urldecode($enc_regs_url), 0,255);
 				$data['enc_image']['url'] = addslashes(abs_url($url, $le_lien));
 				$data['enc_image']['type'] = substr($url, -3);
 			}
 			
-		$enclosure = $data['descriptif'] ;
+		
+		//trouver un flv chez  dailymotion
+		# <media:content url="http://www.dailymotion.com/swf/3ndb67rMbTuLh8E2i" type="application/x-shockwave-flash" duration="520" width="320" height="240"/>
+		//echo $item ;
+		if (preg_match(',(<media:content[^>]*type="video/x-flv".*>),i',
+		$data['item'], $match)) {
+			$go=str_replace('media:content','mediacontent',$match[1]);
+			$data['enc_flash']['url'] = extraire_attribut($go, 'url');
+			$data['enc_flash']['duration'] = extraire_attribut($go, 'duration');
+			$data['enc_flash']['width'] = extraire_attribut($go, 'width');
+			$data['enc_flash']['height'] = extraire_attribut($go, 'height');
+			$data['enc_flash']['type'] = trim(extraire_attribut($go,'type'));
+		}
+		
+		
 		// trouver une image dailymotion
 		// <media:thumbnail url="http://static-02.dailymotion.com/dyn/preview/320x240/2084714.jpg?20070218160100" height="240" width="320"/>			
-			if (preg_match(',(<media:thumbnail[^>]+\/>),i',
+		if (preg_match(',(<media:thumbnail[^>]+\/>),i',
 		$data['item'], $match)) {
 			$go=str_replace('media:thumbnail','mediathumbnail',$match[1]);
 			$data['enc_image']['url'] = extraire_attribut($go, 'url');
@@ -81,10 +74,8 @@
 		
 		/**/
 		
-		//print_r($data['enclosures_all']); die("coucou");
 		PodcastClient_traiter_les_enclosures_rss($data['enclosures_all'],$id_syndic,$le_lien);
-		// mieux vaut preparer les enclosures qu'on veut et passer en une fois, la c n'imp
-		//PodcastClient_traiter_les_enclosures_rss($data['descriptif'],$id_syndic,$le_lien);
+		
 		
 		return func_get_arg(0); # remettre les infos dans le pipeline
 	}
@@ -100,7 +91,7 @@
 		if (!spip_query("SELECT id_syndic, id_syndic_article, id_document FROM spip_documents_syndic")) {
 			spip_log('creation de la table spip_documents_syndic');
 			include_spip('base/create');
- 	        include_spip('base/abstract_sql');
+			include_spip('base/abstract_sql');
             creer_base();
 		}
 	}
@@ -112,20 +103,20 @@
 	//
 	function PodcastClient_traiter_les_enclosures_rss($enclosures,$id_syndic,$le_lien) {
 	spip_log('podcast_client'.$le_lien.'\n\n');
-		//if (sizeof($enclosures) == 0) return false ;
+		if (sizeof($enclosures) == 0) return false ;
 		include_spip('inc/filtres'); # pour extraire_attribut
-	//print_r($enclosures);die("hop");
+	
 		list($id_syndic_article) = spip_fetch_array(spip_query(
 		"SELECT id_syndic_article FROM spip_syndic_articles
 		WHERE id_syndic=$id_syndic AND url='".addslashes($le_lien)."'"), SPIP_NUM);
 	
-		// Attention si cet article est deja vu, ne pas doubler les references // du coup ca efface les mp3 si y'a une image je pense
+		// Attention si cet article est deja vu, ne pas doubler les references 
 		spip_query("DELETE FROM spip_documents_syndic
 		WHERE id_syndic_article=$id_syndic_article");
 		spip_log("efface");
 		// Integrer les enclosures
 		foreach ($enclosures as $enclosure) {
-//echo $enclosure;
+
 			$url = $enclosure['url'] ;
 			$type = $enclosure['type'] ;
 			// href et type sont obligatoires
@@ -142,7 +133,6 @@
 				spip_log("id_type2 : ".$id_type) ;
 				}
 				
-				//echo $id_type; die("coucouman");
 				
 				if (!$id_type) {
 					spip_log("podcast_client: enclosure inconnue ($type) $url");
@@ -172,7 +162,7 @@
 					VALUES ($id_type,'','$url',NOW(),'oui',$taille, 'document', $largeur, $hauteur)");
 					$id_document = spip_insert_id();
 					spip_log("podcast_client: '$url' => id_document=$id_document");
-					//echo "podcast_client: '$url' => id_document=$id_document";
+					
 				}
 	
 				// lier avec l'article syndique
